# ONNX improvements (-62% in full-precision model size, 2.7x faster load and execution, quantizations)