This PR improves the moonshine-onnx package in the following ways:
- Significant reductions in model size / downloads for both tiny and base models by ensuring tied weights are not duplicated, and that the decoders (w/ and w/o past key values) are merged into a single model, without any loss in precision.
    - tiny: **-61.8%** (285MB &rarr; 109MB)
    - base: **-57.6%** (583MB &rarr; 247MB)

- New quantizations (including 4-bit and 8-bit), further reducing the size of the models with minimal differences in output. Note that the q4 quantizations only target MatMul ops, which is why the size is larger than q8 quantization.
    - tiny: 55.1MB at 4-bit quantization, 28MB at 8-bit quantization. Sample outputs:
      ```
      fp32: ['Ever tried ever failed, no matter try again, fail again, fail better.']
      q4: ['Ever tried, ever failed, no matter, try again, fail again, fail better.']
      q8: ['Ever tried. Ever failed. No matter. Try again. Fail again. Fail better.']
      ```
    - base: 98MB at 4-bit quantization, 63MB at 8-bit quantization
      ```
      fp32: ['Ever tried ever failed, no matter try again fail again fail better.']
      q4: ['Ever tried ever failed, no matter try again, fail again, fail better.']
      q8 decoder, fp32 encoder: ['Ever tried ever failed, no matter try again fail again fail better.']
      ```
      (q8 encoder in last case produces poor results)

- Improved loading and execution times, as benchmarked with the following code. Note that these benchmarks do not include downloading time, only loading times (i.e., the models were already downloaded).
    ```python
    import moonshine_onnx as moonshine
    import time
    
    for i in range(10):
        start_time = time.time()
        output = moonshine.transcribe(moonshine.ASSETS_DIR / 'beckett.wav', 'moonshine/tiny')
        end_time = time.time()
    
        print(f"Execution time: {end_time - start_time} seconds")
    ```
    - tiny:
    ![image](https://github.com/user-attachments/assets/ad7c80db-4be2-4fbe-82a7-c4afa92443c3)


    - base:
    ![image](https://github.com/user-attachments/assets/b15e35e1-2da4-4931-9707-dcbcab5598fa)