@xenova 2024-12-15
Absolutely! It's using a custom dev build of [Optimum](https://github.com/huggingface/optimum), which I'll publish soon. It's very similar to the whisper conversion config.

Will do later today 🔥