@keveman 2024-12-16
@xenova Ok to merge this, but will be really grateful for the code to generate the onnx files.