@xenova 2024-12-16
Sure! Just a reminder these are all on dev branches still, and will ready for use when https://github.com/huggingface/transformers/pull/34784 is merged.

Here are the steps to convert:

1. Install the dev branch of Optimum
```
pip install --upgrade git+https://github.com/huggingface/optimum.git@add-moonshine-onnx
```

2. Install moonshine dev branch of transformers:
```
pip install --upgrade git+https://github.com/eustlb/transformers.git@add-moonshine
```

3. Convert the model to ONNX
```
optimum-cli export onnx -m Xenova/moonshine-tiny-hf ./output/
```

---

Note: I've uploaded transformers-compatible versions of the models to my HF account, but I'm happy to move these to your organization, if you'd like. (I can join the org, move, then leave, or you can simply clone the model yourself).
- https://huggingface.co/Xenova/moonshine-tiny-hf
- https://huggingface.co/Xenova/moonshine-base-hf