@keveman 2024-12-16
> Note: I've uploaded transformers-compatible versions of the models to my HF account, but I'm happy to move these to your organization, if you'd like. (I can join the org, move, then leave, or you can simply clone the model yourself).
> 
> * https://huggingface.co/Xenova/moonshine-tiny-hf
> * https://huggingface.co/Xenova/moonshine-base-hf

Sent you an invite to join the usefulsensors org on HF, please move it there.