@xenova 2024-12-16
> Sent you an invite to join the usefulsensors org on HF, please move it there.

Requested to join 👍  (I didn't see an invite, yet. Username is [Xenova](https://huggingface.co/Xenova))