@xenova 2024-12-18
I'm particularly interested in the `_q4` variants, as these are very fast on WebGPU, so doing some evals on that would be great! Also, using the `_quantized` (a.k.a., q8) variant for the encoder can cause some issues, so maybe some hybrid testing?