@xenova 2024-12-18
I'll check out that repo! Regarding file size, remember to deduplicate the the tied weights (as this significantly increases size). 

For example, at fp32, the tiny model is 109.1MB (30.9+78.2 MB):

![image](https://github.com/user-attachments/assets/a1fb4abc-f8b0-4450-afd3-491b7774dc71)

![image](https://github.com/user-attachments/assets/807a49c4-d145-4f91-892b-203d5105b9c8)

and the fp32 base model is 246.8MB (166+80.8MB)

![image](https://github.com/user-attachments/assets/1303ab6c-44b8-4c1f-9330-01e55f572373)

![image](https://github.com/user-attachments/assets/07072e75-6cf9-4d7b-943c-8ccce9f9b54b)